# openapi (4).yaml
openapi: 3.0.1
info:
  title: Book Handler
  description: 'This is a BookHandler server. BookHandler is a platform for sharing and reading books and graphical novels that can be described as online library. This application should be especially useful for book enthusiasts that are looking for rare prints, not available in internet bookstores and on general auction sites and people interested in expanding their personal collections.'
  contact:
    email: 212829@student.pwr.edu.pl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
...
  /genre:
    post:
      tags:
      - genre
      summary: Add new genre to the genre list
      operationId: addGenre
      parameters:
      - name: user
        in: cookie
        description: User info to check genre addition rights
        required: true
        style: form
        explode: true
        schema:
            $ref: '#/components/schemas/User'
      requestBody:
        description: Genre object that needs to be added to the genre list
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicationList'
        required: true
      responses:
        200:
          description: Successfuly added new Genre object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationList'
        409:
          description: Genre name already taken
          content: {}
      x-codegen-request-body-name: body
...
  /genre/{genreName}:
    put:
      tags:
      - genre
      summary: Update genre name
      operationId: updateGenre
      parameters:
      - name: user
        in: cookie
        description: User info to check modification rights
        required: true
        style: form
        explode: true
        schema:
            $ref: '#/components/schemas/User'
      - name: genreName
        in: path
        description: Name of the genre to update
        required: true
        schema:
          type: string
      requestBody:
        description: Genre object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genre'
        required: true
      responses:
        200:
          description: Successfully updated Genre object
          content: {}
        404:
          description: Genre not found
          content: {}
        403:
          description: User does not have the rights to modify Genre objects
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
      - genre
      summary: Delete genre
      operationId: deleteGenre
      parameters:
      - name: user
        description: User info to check genre deletion rights
        in: cookie
        schema:
          $ref: '#/components/schemas/User'
      - name: genreName
        in: path
        description: Name of the genre to delete
        required: true
        schema:
          type: string
      responses:
        200:
          description: Successfully deleted Genre object
          content: {}
        404:
          description: Genre not found
          content: {}
        403:
          description: User does not have the rights to delete genre
          content: {}
components:
  schemas:
    Publication:
      required:
      - title
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        releaseYear:
          type: integer
          format: int32
        genre:
          type: array
          items: 
            $ref: '#/components/schemas/Genre'
        dateAdded:
          type: string
          format: date-time
        rating:
          type: number
          format: float
        photoUrls:
          type: array
          xml:
            name: photoUrl
            wrapped: true
          items:
            type: string
        addedBy:
          $ref: '#/components/schemas/User'
    Genre:
      type: object
      properties:
        name:
          type: string
    User:
      required:
      - username
      - email
      - password
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        userRole:
          type: string
          description: User Status
          enum: 
            - admin
            - member
        dateAdded:
          type: string
          format: date-time
    PublicationList:
      required:
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        createdBy:
          $ref: '#/components/schemas/User'
        publications:
          type: array
          items: 
            $ref: '#/components/schemas/Publication'
        dateAdded:
          type: string
          format: date-time